requirements.txt
langchain-google-genai
langchain-ollama
langchain
langchain-community
pydantic
faiss-cpu
sentence-transformers
pandas
python-dotenv
pyyaml
typer
rich